requirements.txt
openai==0.28
langchain
faiss-cpu
streamlit
chromadb
python-dotenv
pypdf
pymongo
tiktoken
